+++ CHANGELOG.md
@@ -2,6 +2,6 @@
 
 ## 0.6.0
 
+- Add stream parsers for PICA Plain, PICA Normalized and Annotated PICA
 - Support parsing empty subfield values
 - Set empty/zero occurrence to empty string instead of null
-- Add stream parsers